+++ src/discord/index.ts
@@ -2,6 +2,7 @@ import { Client, Events, GatewayIntentBits, Collection, Interaction } from 'disc
 import { token } from 'config.json';
 import { collectCommands } from '@/discord/collectFiles';
 import type { Command } from '@/types';
+import executeCommands from '@/discord/events/executeCommands';
 
 
 (async () => {
@@ -21,30 +22,7 @@ import type { Command } from '@/types';
     console.log(`Ready! Logged in as ${readyClient.user.tag}`);
   });
   
-  client.on(Events.InteractionCreate, async (interaction: Interaction) => {
-    if (!interaction.isChatInputCommand()) return;
-    const command = interaction.client.commands.get(interaction.commandName)
-  
-    if (!command) {
-      console.error(
-        `No command matching ${interaction.commandName} was found.`,
-      );
-      return;
-    }
-    try {
-      command.execute(interaction);
-      console.log("executed command: ", interaction.commandName);
-    } catch (error) {
-      console.error(`Error executing ${interaction.commandName}`);
-      console.error(error);
-  
-      if(interaction.replied || interaction.deferred) {
-        await interaction.followUp({ content: 'There was an error while executing this command!', ephemeral: true });
-      } else {
-        await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
-      }
-    }
-  });
+  client.on(Events.InteractionCreate, executeCommands);
   
   client.login(token);
 })();
